Mega motor size increase and Reformat of robot loop structure
For the Arduino Mega the motor array is now int16_t to allow values
greater then (-127, 126). The motor values will now be checked and set
to the max min value of a byte then casted to int8_t for serial
communication.

Both files main loop is now robotLoop() which is called from the main
Arduino loop().
The method update() is now used to  update sensors and to update serial.

An example of a suggest design to control the robot updates is in
comments.